🐛 : Fix RMSE Loss Accumulation
Made significant changes to this file. It is now A LOT nicer to work with. Comments are very clear. We're also ensuring that we train on MSE loss, with RMSE loss used correctly for validation. We are not accumulating RMSE as we go over each snapshot; it's calculated at the end of each epoch instead.