🔧 : Slight Tweaks to Training
Changed from RMSE loss to MSE loss as that's probably a better metric to use for training. RMSE loss will still be used for validation. Also changed the shuffling mechanic to remove the specific seeding, as that's just unnecessary. Also contains the graphs from a very successful new training experiment.