🍌 : Deprecate Subset Ratio
The option to use only a fraction of the original dataset is being sunset to make way for a new world where the entire training set is used every epoch. The trade-off doesn't seem worthwhile. References to the subset ratio will be removed from the loss plot function in this update. Deprecation is referenced in the documentation.